+++ gaussian/run_gaussian/runGau04_hokusai.sh
@@ -1,34 +1,39 @@
 #!/bin/bash
 
 # -----------------------------------------------
-QMINP=$1
-QMOUT=$2
-NSTEP=$3
-MOL=${QMINP%.*}
-
-initialchk='../initial.chk'
-
-# -----------------------------------------------
-# Initial MO
-#
-# In the following, the initial MO is copied only 
-# for the 1st step in MD.
-#
-if [ $NSTEP -eq 0 ] && [ -e ${initialchk} ]; then
-  cp ${initialchk} gaussian.chk
-fi
-
-# -----------------------------------------------
 # Settings for Gaussian16
 #
 module load gaussian
 GAUSS_EXEBIN=g16
 
+# --- Set the path for a scratch folder ---
+scratch=./g16scratch
+
+# (optional) 
+# --- Set a chkpoint file to read initial MOs ---
+#initialchk='../initial.chk'
+
+# -----------------------------------------------
+
+QMINP=$1
+QMOUT=$2
+NSTEP=$3
+MOL=${QMINP%.*}
+
+# -----------------------------------------------
+# Initial MO
+#
+# The initial MO is copied only for the 1st step 
+# in MD.
+#
+if [ $NSTEP -eq 0 ] && [ -n "${initialchk}" ] && [ -e ${initialchk} ]; then
+  cp ${initialchk} gaussian.chk
+fi
+
 # -----------------------------------------------
 # Scratch folder settings
 #
-TIME=$(date '+%N')
-export GAUSS_SCRDIR=./g16scratch/$MOL.$TIME.$$
+export GAUSS_SCRDIR=$scratch/$(mktemp -u $MOL.XXXX)
 mkdir -p ${GAUSS_SCRDIR}
 
 # -----------------------------------------------
